+++ .gitignore
@@ -1,2 +1,3 @@
 saved_weights.pt
+__pycache__/*
 __pycache__
