+++ Alternative method/config.py
@@ -1,0 +1,22 @@
+# Parameters
+# run_on_cpu
+PARAM1 = "true"
+# training dataset
+PARAM2 = "7000_training_dataset.csv"
+# test dataset
+PARAM3 = "500_test_dataset.csv"
+# this is variable for spliting training dataset into train and validation
+# test data set is already spited to other file
+PARAM4 = 0.2
+# numers of epochs
+PARAM5 = 5
+# clean stopwords
+PARAM6= 1
+# batch_size
+PARAM7 = 16
+
+
+# 1 = train, 0 = no train
+PARAM10 = 0
+# 1 = get predictons, 0 = don't get predictions
+PARAM11 = 1
